Fix the bug in compare value
GCC fill zero in padding bits in struct and union are undifined behavior to C std, and clang not do this. So if we use clang to compile code, we will see the error result of comparing isns_value_t.
So I fill zero in initialization of isns_value_t.