Testing: update testing system
Udate the README file and rename to README.md, so it can use
markdown. Also update the README file in tests, renaming it
to README.md as well.

Update the tests so that they can be run from other directories
than the tests subdirectory.

Update the meson build system so that running the unit tests
works correctly from the source directory.

Added a "-l"/"--list" option to the test script that can list
all the unit tests and their subtests.